Change of Extension and Capital Letters
-Change from Example Extension to Agilethought
-Change of capital letters to packages
