--- at-sce-api/src/main/java/com/example/atsceapi/AtSceApiApplication.java
@@ -1,13 +1,0 @@
-package com.example.atsceapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AtSceApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AtSceApiApplication.class, args);
-	}
-
-}
