--- at-sce-api/src/main/java/com/example/atsceapi/Controller/UserController.java
@@ -1,5 +1,0 @@
-package com.example.atsceapi.Controller;
-
-public class UserController {
-
-}
