--- at-sce-api/src/main/java/com/example/atsceapi/Repository/UserRepository.java
@@ -1,5 +1,0 @@
-package com.example.atsceapi.Repository;
-
-public interface UserRepository {
-
-}
