--- at-sce-api/src/main/java/com/example/atsceapi/Service/UserService.java
@@ -1,5 +1,0 @@
-package com.example.atsceapi.Service;
-
-public interface UserService {
-
-}
