--- at-sce-api/src/test/java/com/example/atsceapi/AtSceApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.atsceapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AtSceApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
