+++ src/main/java/com/agilethought/atsceapi/AtSceApiApplication.java
@@ -1,0 +1,13 @@
+package com.agilethought.atsceapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AtSceApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AtSceApiApplication.class, args);
+	}
+
+}
