+++ src/main/java/com/agilethought/atsceapi/controller/UserController.java
@@ -1,0 +1,5 @@
+package com.agilethought.atsceapi.controller;
+
+public class UserController {
+
+}
