+++ src/main/java/com/agilethought/atsceapi/repository/UserRepository.java
@@ -1,0 +1,5 @@
+package com.agilethought.atsceapi.repository;
+
+public interface UserRepository {
+
+}
