+++ src/main/java/com/agilethought/atsceapi/service/UserService.java
@@ -1,0 +1,5 @@
+package com.agilethought.atsceapi.service;
+
+public interface UserService {
+
+}
