+++ src/main/java/com/agilethought/atsceapi/service/UserServiceImpl.java
@@ -1,0 +1,5 @@
+package com.agilethought.atsceapi.service;
+
+public class UserServiceImpl {
+
+}
