+++ src/test/java/com/agilethought/atsceapi/AtSceApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.agilethought.atsceapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AtSceApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
